requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
sqlmodel==0.0.14

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Firebase
firebase-admin==6.3.0

# SendGrid
sendgrid==6.11.0

# Razorpay
razorpay==1.4.1

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0

# Other
httpx==0.25.2
redis==5.0.1
celery==5.3.4
transactions
boto3
fpdf
slowapi
pytest
httpx
paho-mqtt
